+++ firebase_rules_generator/test_project/lib/firebase/namespace/path/input.dart
@@ -6,13 +6,11 @@ bool test(RulesPath other) {
   final a = rules.path('/path/to/resource'.rules());
   final b = other['fieldname'];
   final c = other[0];
-  final d = other.bind(
-    {
-      'foo': 'something',
-      'bar': 'another',
-    }.rules(),
-  );
+  final d = other.bind({'foo': 'something', 'bar': 'another'}.rules());
   final e = rules.path('/path/to/resource'.rules(), database: 'default');
+  final f = rules
+      .path('/path/to/resource'.rules())
+      .bind({'foo': 'something', 'bar': 'another'}.rules());
   return true;
 }