+++ firebase_rules_generator/CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fixes issue with raw matching
 - Raws are now completely ignored during sanitization
 - Improvements to string interpolation handling
+- `rules.path('/path/to/resource')` now translates to `/path/to/resource`
 
 ## 0.0.1
 - Initial release